move snapshot incremental logic to dataplane
