+++ cloud/disk_manager/internal/pkg/dataplane/create_snapshot_from_snapshot_task.go
@@ -100,7 +100,8 @@ func (t *createSnapshotFromSnapshotTask) Cancel(
 	execCtx tasks.ExecutionContext,
 ) error {
 
-	return t.storage.DeletingSnapshot(ctx, t.request.DstSnapshotId)
+	_, err := t.storage.DeletingSnapshot(ctx, t.request.DstSnapshotId, execCtx.GetTaskID())
+	return err
 }
 
 func (t *createSnapshotFromSnapshotTask) GetMetadata(
