+++ cloud/disk_manager/internal/pkg/dataplane/create_snapshot_from_url_task.go
@@ -176,7 +176,8 @@ func (t *createSnapshotFromURLTask) Cancel(
 	execCtx tasks.ExecutionContext,
 ) error {
 
-	return t.storage.DeletingSnapshot(ctx, t.request.DstSnapshotId)
+	_, err := t.storage.DeletingSnapshot(ctx, t.request.DstSnapshotId, execCtx.GetTaskID())
+	return err
 }
 
 func (t *createSnapshotFromURLTask) GetMetadata(
