+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/common.go
@@ -48,13 +48,16 @@ type snapshotState struct {
 	zoneID            string
 	diskID            string
 	checkpointID      string
+	createTaskID      string
 	creatingAt        time.Time
 	createdAt         time.Time
 	deletingAt        time.Time
 	baseSnapshotID    string
+	baseCheckpointID  string
 	size              uint64
 	storageSize       uint64
 	chunkCount        uint32
+	lockTaskID        string
 	encryptionMode    uint32
 	encryptionKeyHash []byte
 	status            snapshotStatus
@@ -62,9 +65,18 @@ type snapshotState struct {
 
 func (s *snapshotState) toSnapshotMeta() *SnapshotMeta {
 	return &SnapshotMeta{
-		Size:        s.size,
-		StorageSize: s.storageSize,
-		ChunkCount:  s.chunkCount,
+		ID: s.id,
+		Disk: &types.Disk{
+			ZoneId: s.zoneID,
+			DiskId: s.diskID,
+		},
+		CheckpointID:     s.checkpointID,
+		CreateTaskID:     s.createTaskID,
+		BaseSnapshotID:   s.baseSnapshotID,
+		BaseCheckpointID: s.baseCheckpointID,
+		Size:             s.size,
+		StorageSize:      s.storageSize,
+		ChunkCount:       s.chunkCount,
 		Encryption: &types.EncryptionDesc{
 			Mode: types.EncryptionMode(s.encryptionMode),
 			Key: &types.EncryptionDesc_KeyHash{
@@ -81,13 +93,16 @@ func (s *snapshotState) structValue() persistence.Value {
 		persistence.StructFieldValue("zone_id", persistence.UTF8Value(s.zoneID)),
 		persistence.StructFieldValue("disk_id", persistence.UTF8Value(s.diskID)),
 		persistence.StructFieldValue("checkpoint_id", persistence.UTF8Value(s.checkpointID)),
+		persistence.StructFieldValue("create_task_id", persistence.UTF8Value(s.createTaskID)),
 		persistence.StructFieldValue("creating_at", persistence.TimestampValue(s.creatingAt)),
 		persistence.StructFieldValue("created_at", persistence.TimestampValue(s.createdAt)),
 		persistence.StructFieldValue("deleting_at", persistence.TimestampValue(s.deletingAt)),
 		persistence.StructFieldValue("base_snapshot_id", persistence.UTF8Value(s.baseSnapshotID)),
+		persistence.StructFieldValue("base_checkpoint_id", persistence.UTF8Value(s.baseCheckpointID)),
 		persistence.StructFieldValue("size", persistence.Uint64Value(s.size)),
 		persistence.StructFieldValue("storage_size", persistence.Uint64Value(s.storageSize)),
 		persistence.StructFieldValue("chunk_count", persistence.Uint32Value(s.chunkCount)),
+		persistence.StructFieldValue("lock_task_id", persistence.UTF8Value(s.lockTaskID)),
 		persistence.StructFieldValue("encryption_mode", persistence.Uint32Value(s.encryptionMode)),
 		persistence.StructFieldValue("encryption_keyhash", persistence.StringValue(s.encryptionKeyHash)),
 		persistence.StructFieldValue("status", persistence.Int64Value(int64(s.status))),
@@ -100,13 +115,16 @@ func scanSnapshotState(res persistence.Result) (state snapshotState, err error) 
 		persistence.OptionalWithDefault("zone_id", &state.zoneID),
 		persistence.OptionalWithDefault("disk_id", &state.diskID),
 		persistence.OptionalWithDefault("checkpoint_id", &state.checkpointID),
+		persistence.OptionalWithDefault("create_task_id", &state.createTaskID),
 		persistence.OptionalWithDefault("creating_at", &state.creatingAt),
 		persistence.OptionalWithDefault("created_at", &state.createdAt),
 		persistence.OptionalWithDefault("deleting_at", &state.deletingAt),
 		persistence.OptionalWithDefault("base_snapshot_id", &state.baseSnapshotID),
+		persistence.OptionalWithDefault("base_checkpoint_id", &state.baseCheckpointID),
 		persistence.OptionalWithDefault("size", &state.size),
 		persistence.OptionalWithDefault("storage_size", &state.storageSize),
 		persistence.OptionalWithDefault("chunk_count", &state.chunkCount),
+		persistence.OptionalWithDefault("lock_task_id", &state.lockTaskID),
 		persistence.OptionalWithDefault("encryption_mode", &state.encryptionMode),
 		persistence.OptionalWithDefault("encryption_keyhash", &state.encryptionKeyHash),
 		persistence.OptionalWithDefault("status", &state.status),
@@ -143,13 +161,16 @@ func snapshotStateStructTypeString() string {
 		zone_id: Utf8,
 		disk_id: Utf8,
 		checkpoint_id: Utf8,
+		create_task_id: Utf8,
 		creating_at: Timestamp,
 		created_at: Timestamp,
 		deleting_at: Timestamp,
 		base_snapshot_id: Utf8,
+		base_checkpoint_id: Utf8,
 		size: Uint64,
 		storage_size: Uint64,
 		chunk_count: Uint32,
+		lock_task_id: Utf8,
 		encryption_mode: Uint32,
 		encryption_keyhash: String,
 		status: Int64>`
