+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/storage.go
@@ -12,10 +12,12 @@ import (
 ////////////////////////////////////////////////////////////////////////////////
 
 type SnapshotMeta struct {
-	ID             string
-	Disk           *types.Disk
-	CheckpointID   string
-	BaseSnapshotID string
+	ID               string
+	Disk             *types.Disk
+	CheckpointID     string
+	CreateTaskID     string
+	BaseSnapshotID   string
+	BaseCheckpointID string
 	// Snapshot virtual size, i.e. the minimum amount of disk space needed to restore.
 	Size uint64
 	// Snapshot real size, i.e. the amount of disk space occupied in storage.
@@ -50,7 +52,11 @@ type Storage interface {
 		encryption *types.EncryptionDesc,
 	) error
 
-	DeletingSnapshot(ctx context.Context, snapshotID string) error
+	DeletingSnapshot(
+		ctx context.Context,
+		snapshotID string,
+		taskID string,
+	) (*SnapshotMeta, error)
 
 	GetSnapshotsToDelete(
 		ctx context.Context,
@@ -118,4 +124,12 @@ type Storage interface {
 		zoneID string,
 		diskID string,
 	) error
+
+	LockSnapshot(
+		ctx context.Context,
+		snapshotID string,
+		lockTaskID string,
+	) (locked bool, err error)
+
+	UnlockSnapshot(ctx context.Context, snapshotID string, lockTaskID string) error
 }
