+++ cloud/disk_manager/internal/pkg/services/snapshots/delete_snapshot_task.go
@@ -13,7 +13,6 @@ import (
 	"github.com/ydb-platform/nbs/cloud/tasks"
 	"github.com/ydb-platform/nbs/cloud/tasks/errors"
 	"github.com/ydb-platform/nbs/cloud/tasks/headers"
-	"github.com/ydb-platform/nbs/cloud/tasks/logging"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -62,24 +61,6 @@ func (t *deleteSnapshotTask) deleteSnapshot(
 		return errors.NewNonCancellableErrorf(
 			"id %v is not accepted",
 			t.request.SnapshotId,
-		)
-	}
-
-	// NBS-3535.
-	if snapshotMeta.UseDataplaneTasks && len(snapshotMeta.BaseSnapshotID) != 0 {
-		err := t.storage.UnlockSnapshot(
-			ctx,
-			snapshotMeta.BaseSnapshotID,
-			snapshotMeta.CreateTaskID,
-		)
-		if err != nil {
-			return err
-		}
-
-		logging.Info(
-			ctx,
-			"Successfully unlocked snapshot with id %v",
-			snapshotMeta.BaseSnapshotID,
 		)
 	}
